[Rust] Rename algorithm submodule to visitor
